feat(client): move Linux packet capture into a privilege-separated helper (#726)
Server detection on Linux sniffs UDP through an AF_PACKET socket that needs
CAP_NET_RAW. Running that in-process meant the whole Tauri/webkit GUI would have
to hold the capability. Split the packet capture into a tiny betterfleet-netcap
binary that carries CAP_NET_RAW on its own, so the GUI stays unprivileged.

The capture core (FlowStat, the aggregator and ranking, parse_game_flow,
is_plausible_sot_port and the AF_PACKET loop) moves into a new Tauri-free crate,
better_fleet_netcap, re-exported as better_fleet::capture so the app keeps a
single copy of the code. The helper links only that crate, socket2, etherparse
and serde_json, never Tauri.

On Linux capture_flows now spawns the helper first (next to the executable, then
on PATH), reads back the ranked flows as JSON, and falls back to an in-process
capture when the helper is missing or cannot open the socket, so a developer who
ran setcap on the GUI binary still works. Both live detection and the Help-tab
diagnostic go through capture_flows, so both benefit. Windows and macOS capture
paths are unchanged.
