+++ webapp/src-tauri/src/diagnostics.rs
@@ -10,32 +10,22 @@ use std::net::{IpAddr, SocketAddr, ToSocketAddrs};
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, Instant};
 
-use etherparse::PacketHeaders;
 use log::error;
-use serde::{Deserialize, Serialize};
-use tokio::net::UdpSocket;
-
-use crate::fetch_informations::{get_hostname, is_plausible_sot_port};
-#[cfg(windows)]
-use crate::fetch_informations::create_raw_socket;
 #[cfg(target_os = "linux")]
-use socket2::{Domain, Protocol, Socket, Type};
+use log::info;
+use serde::Serialize;
 
-/// One observed UDP conversation between a game-owned local port and a remote peer.
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
-pub struct FlowStat {
-    pub local_port: u16,
-    pub remote_ip: String,
-    pub remote_port: u16,
-    pub packets: u32,
-    pub bytes: u64,
-    pub inbound: u32,
-    pub outbound: u32,
-    /// Remote port sits in the range Sea of Thieves game servers use.
-    pub plausible_sot_port: bool,
-    pub first_seen_ms: u64,
-    pub last_seen_ms: u64,
-}
+// The capture primitives now live in the Tauri-free better_fleet_netcap crate (#726), re-exported as
+// better_fleet::capture by src/lib.rs. FlowStat is used on every platform; the aggregator and the raw
+// packet parser are only needed by the Windows in-process sniff below.
+use better_fleet::capture::FlowStat;
+#[cfg(windows)]
+use better_fleet::capture::{parse_game_flow, FlowAggregator};
+
+#[cfg(windows)]
+use crate::fetch_informations::{create_raw_socket, get_hostname};
+#[cfg(windows)]
+use tokio::net::UdpSocket;
 
 /// The full result of a diagnostic capture, ready to be serialized and shared.
 #[derive(Serialize, Clone, Debug)]
@@ -57,96 +47,6 @@ pub struct DiagnosticReport {
     pub top_candidates: Vec<FlowStat>,
     /// Every observed flow, ranked by volume.
     pub flows: Vec<FlowStat>,
-}
-
-/// Aggregates observed UDP packets into per-flow statistics. Deliberately free of
-/// any I/O so the aggregation and ranking can be unit-tested deterministically.
-#[derive(Default)]
-pub struct FlowAggregator {
-    map: HashMap<(u16, String, u16), FlowStat>,
-}
-
-impl FlowAggregator {
-    /// Record one packet on the flow (local_port <-> remote_ip:remote_port).
-    /// `inbound` is true when the game is the destination, false when it's the source.
-    pub fn observe(
-        &mut self,
-        local_port: u16,
-        remote_ip: &str,
-        remote_port: u16,
-        len: usize,
-        inbound: bool,
-        t_ms: u64,
-    ) {
-        let entry = self
-            .map
-            .entry((local_port, remote_ip.to_string(), remote_port))
-            .or_insert_with(|| FlowStat {
-                local_port,
-                remote_ip: remote_ip.to_string(),
-                remote_port,
-                packets: 0,
-                bytes: 0,
-                inbound: 0,
-                outbound: 0,
-                plausible_sot_port: is_plausible_sot_port(remote_port),
-                first_seen_ms: t_ms,
-                last_seen_ms: t_ms,
-            });
-        entry.packets += 1;
-        entry.bytes += len as u64;
-        if inbound {
-            entry.inbound += 1;
-        } else {
-            entry.outbound += 1;
-        }
-        entry.last_seen_ms = t_ms;
-    }
-
-    /// Drains the aggregated flows, sorted by packet volume (desc), then bytes
-    /// (desc), then local port (asc) for a stable order.
-    pub fn take_sorted_flows(&mut self) -> Vec<FlowStat> {
-        let mut flows: Vec<FlowStat> = std::mem::take(&mut self.map).into_values().collect();
-        flows.sort_by(|a, b| {
-            b.packets
-                .cmp(&a.packets)
-                .then(b.bytes.cmp(&a.bytes))
-                .then(a.local_port.cmp(&b.local_port))
-        });
-        flows
-    }
-}
-
-/// Parses one raw IP packet and, when it belongs to one of the game's UDP ports,
-/// returns (game_local_port, remote_ip, remote_port, inbound).
-fn parse_game_flow(bytes: &[u8], game_ports: &HashSet<u16>) -> Option<(u16, String, u16, bool)> {
-    let packet = PacketHeaders::from_ip_slice(bytes).ok()?;
-
-    let (source_ip, destination_ip) = match packet.net? {
-        etherparse::NetHeaders::Ipv4(header, _) => (
-            std::net::Ipv4Addr::from(header.source).to_string(),
-            std::net::Ipv4Addr::from(header.destination).to_string(),
-        ),
-        etherparse::NetHeaders::Ipv6(header, _) => (
-            std::net::Ipv6Addr::from(header.source).to_string(),
-            std::net::Ipv6Addr::from(header.destination).to_string(),
-        ),
-    };
-
-    let (source_port, destination_port) = match packet.transport? {
-        etherparse::TransportHeader::Udp(header) => (header.source_port, header.destination_port),
-        _ => return None,
-    };
-
-    if game_ports.contains(&source_port) {
-        // Game is the source -> outbound
-        Some((source_port, destination_ip, destination_port, false))
-    } else if game_ports.contains(&destination_port) {
-        // Game is the destination -> inbound
-        Some((destination_port, source_ip, source_port, true))
-    } else {
-        None
-    }
 }
 
 /// Sniffs a single local interface for `duration`, feeding every game-owned UDP
@@ -227,85 +127,119 @@ pub async fn capture_flows(game_ports: Vec<u16>, window: Duration) -> Vec<FlowSt
     flows
 }
 
-/// Linux raw-capture backend (#725). A single `AF_PACKET`/`SOCK_DGRAM` socket sees every IP packet
-/// crossing the host in both directions across all interfaces - no per-interface fan-out like the
-/// Windows path - and the kernel strips the link-layer header, so each datagram arrives
-/// network-layer-first and feeds the SAME `parse_game_flow` + `FlowAggregator` the Windows sniff
-/// does. Ranking-by-volume is therefore shared verbatim.
+/// Linux raw-capture backend (#725, #726). Server detection needs an `AF_PACKET` socket, which needs
+/// `CAP_NET_RAW`. To keep the Tauri GUI unprivileged, the capture is isolated in a tiny helper binary
+/// (`betterfleet-netcap`) that carries the capability on its own: the GUI spawns it with the game
+/// ports and window, and reads back the ranked flows as JSON. If the helper is unavailable or fails -
+/// for instance a dev build with no capability granted, or a packaging that shipped without it - we
+/// fall back to capturing in-process, which still works for a developer who ran `setcap` on the GUI
+/// binary itself. Either path feeds the same [`better_fleet_netcap`] ranking, so live detection and
+/// the diagnostic behave identically whichever one ran. A total failure yields no flows, so detection
+/// degrades to "no server" rather than crashing.
 ///
-/// Needs `CAP_NET_RAW`: in dev, run as root or apply `sudo setcap cap_net_raw+ep` to the binary; the
-/// productionised capability grant is issue #726 (out of scope here). A permission failure is logged
-/// and simply yields no flows, so detection degrades to "no server" rather than crashing.
+/// Shared by the diagnostic report and by live detection (both call `capture_flows`), so both benefit
+/// from the privilege-separated helper without either knowing which path served the flows.
 #[cfg(target_os = "linux")]
 pub async fn capture_flows(game_ports: Vec<u16>, window: Duration) -> Vec<FlowStat> {
-    let port_set: HashSet<u16> = game_ports.into_iter().collect();
-    // AF_PACKET recv blocks; run the capture on a blocking thread so the detection task's runtime is
-    // never stalled, and await its result - mirroring how the Windows path awaits its sniff tasks.
-    match tokio::task::spawn_blocking(move || capture_af_packet(port_set, window)).await {
+    if let Some(flows) = capture_via_helper(&game_ports, window).await {
+        return flows;
+    }
+    // The helper was missing or could not capture; fall back to an in-process capture, which needs
+    // CAP_NET_RAW on the GUI binary itself. AF_PACKET recv blocks, so run it on a blocking thread and
+    // await the result, mirroring how the Windows path awaits its sniff tasks.
+    info!("[capture] falling back to in-process AF_PACKET capture (betterfleet-netcap unavailable)");
+    match tokio::task::spawn_blocking(move || better_fleet::capture::run_capture(game_ports, window))
+        .await
+    {
         Ok(flows) => flows,
         Err(e) => {
-            error!("[capture] AF_PACKET capture thread failed: {e}");
+            error!("[capture] in-process AF_PACKET capture thread failed: {e}");
             Vec::new()
         }
     }
 }
 
-/// Blocking AF_PACKET capture loop, split out of [`capture_flows`] so it can run under
-/// `spawn_blocking`.
+/// Runs the `betterfleet-netcap` helper for one capture window and returns the flows it printed, or
+/// `None` if the helper could not be used - missing, failed to spawn, exited non-zero (for instance
+/// it lacks `CAP_NET_RAW`), or printed output we could not parse - so the caller falls back to
+/// in-process capture. The blocking process wait runs on a blocking thread so the detection runtime
+/// is never stalled.
 #[cfg(target_os = "linux")]
-fn capture_af_packet(game_ports: HashSet<u16>, window: Duration) -> Vec<FlowStat> {
-    use std::mem::MaybeUninit;
+async fn capture_via_helper(game_ports: &[u16], window: Duration) -> Option<Vec<FlowStat>> {
+    let helper = locate_capture_helper();
+    let ports_arg = game_ports
+        .iter()
+        .map(|port| port.to_string())
+        .collect::<Vec<_>>()
+        .join(",");
+    // The helper takes a whole-seconds window; keep at least 1s so a sub-second window still captures.
+    let window_arg = window.as_secs().max(1).to_string();
 
-    // ETH_P_ALL, in network byte order because AF_PACKET takes its protocol big-endian (== htons).
-    // SOCK_DGRAM (not SOCK_RAW) tells the kernel to strip the link-layer header, so recv() yields the
-    // IP packet directly - exactly what parse_game_flow (from_ip_slice) expects, as on Windows.
-    const ETH_P_ALL: u16 = 0x0003;
-    let protocol = Protocol::from(i32::from(ETH_P_ALL.to_be()));
-    let socket = match Socket::new(Domain::PACKET, Type::DGRAM, Some(protocol)) {
-        Ok(socket) => socket,
+    let output = tokio::task::spawn_blocking(move || {
+        std::process::Command::new(&helper)
+            .arg(&ports_arg)
+            .arg(&window_arg)
+            .stdout(std::process::Stdio::piped())
+            .stderr(std::process::Stdio::piped())
+            .output()
+    })
+    .await;
+
+    let output = match output {
+        Ok(Ok(output)) => output,
+        Ok(Err(e)) => {
+            if e.kind() == std::io::ErrorKind::NotFound {
+                info!("[capture] betterfleet-netcap not found beside the executable or on PATH");
+            } else {
+                error!("[capture] could not spawn betterfleet-netcap: {e}");
+            }
+            return None;
+        }
         Err(e) => {
-            error!(
-                "[capture] AF_PACKET socket failed: {e} — needs CAP_NET_RAW (run as root or \
-                 `sudo setcap cap_net_raw+ep` in dev; productionised in #726)"
-            );
-            return Vec::new();
+            error!("[capture] betterfleet-netcap wait thread failed: {e}");
+            return None;
         }
     };
-    // A read timeout lets the loop honour the window deadline even when no packet arrives.
-    if let Err(e) = socket.set_read_timeout(Some(Duration::from_millis(200))) {
-        error!("[capture] AF_PACKET set_read_timeout failed: {e}");
-        return Vec::new();
+
+    if !output.status.success() {
+        let stderr = String::from_utf8_lossy(&output.stderr);
+        error!(
+            "[capture] betterfleet-netcap exited with {}: {}",
+            output.status,
+            stderr.trim()
+        );
+        return None;
     }
 
-    let mut aggregator = FlowAggregator::default();
-    // A full IP datagram fits in 64 KiB; one buffer, reused per packet.
-    let mut buf = [MaybeUninit::<u8>::uninit(); 65535];
-    let start = Instant::now();
-    while start.elapsed() < window {
-        match socket.recv(&mut buf) {
-            Ok(0) => {}
-            Ok(len) => {
-                // SAFETY: recv reports `len` bytes written, so the first `len` bytes are initialised;
-                // MaybeUninit<u8> and u8 share layout, so viewing them as `&[u8]` is sound.
-                let data = unsafe { std::slice::from_raw_parts(buf.as_ptr() as *const u8, len) };
-                if let Some((local_port, remote_ip, remote_port, inbound)) =
-                    parse_game_flow(data, &game_ports)
-                {
-                    let t_ms = start.elapsed().as_millis() as u64;
-                    aggregator.observe(local_port, &remote_ip, remote_port, len, inbound, t_ms);
-                }
-            }
-            // Read timeout (SO_RCVTIMEO surfaces as WouldBlock/TimedOut): re-check the deadline.
-            Err(e)
-                if e.kind() == std::io::ErrorKind::WouldBlock
-                    || e.kind() == std::io::ErrorKind::TimedOut => {}
-            Err(e) => {
-                error!("[capture] AF_PACKET recv error: {e}");
-                break;
+    match serde_json::from_slice::<Vec<FlowStat>>(&output.stdout) {
+        Ok(flows) => {
+            info!("[capture] betterfleet-netcap returned {} flow(s)", flows.len());
+            Some(flows)
+        }
+        Err(e) => {
+            error!("[capture] could not parse betterfleet-netcap output: {e}");
+            None
+        }
+    }
+}
+
+/// Finds the `betterfleet-netcap` helper: next to the running executable first (where the packaged
+/// helper is installed alongside the GUI), otherwise a bare name for the OS to resolve on `PATH` (a
+/// dev convenience - `cargo build` drops it in the same target dir). Returning a bare name means a
+/// genuinely missing helper surfaces as a `NotFound` spawn error, which the caller treats as "fall
+/// back to in-process".
+#[cfg(target_os = "linux")]
+fn locate_capture_helper() -> std::path::PathBuf {
+    const HELPER: &str = "betterfleet-netcap";
+    if let Ok(exe) = std::env::current_exe() {
+        if let Some(dir) = exe.parent() {
+            let candidate = dir.join(HELPER);
+            if candidate.is_file() {
+                return candidate;
             }
         }
     }
-    aggregator.take_sorted_flows()
+    std::path::PathBuf::from(HELPER)
 }
 
 /// macOS (and any other non-Windows, non-Linux target): no capture backend yet, so detection reports
@@ -438,58 +372,10 @@ pub async fn run_diagnostic(
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn busiest_plausible_flow_ranks_first() {
-        let mut agg = FlowAggregator::default();
-        // Two sparse SDR-like relay flows (Steam-owned peers, non-SoT ports).
-        agg.observe(50001, "162.254.1.1", 27017, 60, true, 10);
-        agg.observe(50002, "162.254.1.2", 27018, 60, true, 20);
-        // The busy game-server flow (plausible SoT remote port), sustained traffic.
-        for i in 0..20u64 {
-            agg.observe(59639, "20.1.2.3", 35000, 120, i % 2 == 0, 100 + i);
-        }
-
-        let flows = agg.take_sorted_flows();
-        assert_eq!(flows.len(), 3);
-
-        // Highest-volume flow wins and is flagged as a plausible server.
-        assert_eq!(flows[0].local_port, 59639);
-        assert_eq!(flows[0].packets, 20);
-        assert_eq!(flows[0].bytes, 20 * 120);
-        assert!(flows[0].plausible_sot_port);
-        assert_eq!(flows[0].inbound, 10);
-        assert_eq!(flows[0].outbound, 10);
-
-        // The sparse relay flows are not flagged as SoT candidates.
-        assert!(!flows[1].plausible_sot_port);
-        assert!(!flows[2].plausible_sot_port);
-    }
-
-    #[test]
-    fn packets_on_the_same_flow_accumulate() {
-        let mut agg = FlowAggregator::default();
-        agg.observe(59639, "20.1.2.3", 35000, 100, false, 0);
-        agg.observe(59639, "20.1.2.3", 35000, 140, true, 5);
-
-        let flows = agg.take_sorted_flows();
-        assert_eq!(flows.len(), 1);
-        assert_eq!(flows[0].packets, 2);
-        assert_eq!(flows[0].bytes, 240);
-        assert_eq!(flows[0].inbound, 1);
-        assert_eq!(flows[0].outbound, 1);
-        assert_eq!(flows[0].first_seen_ms, 0);
-        assert_eq!(flows[0].last_seen_ms, 5);
-    }
-
-    #[test]
-    fn different_remotes_on_one_local_port_are_distinct_flows() {
-        // A single local port talking to two different peers must not be merged.
-        let mut agg = FlowAggregator::default();
-        agg.observe(60000, "1.1.1.1", 35000, 50, true, 0);
-        agg.observe(60000, "2.2.2.2", 35001, 50, true, 1);
-        assert_eq!(agg.take_sorted_flows().len(), 2);
-    }
+    // is_plausible_sot_port moved to the capture crate with the aggregator (#726); the ranking/pick
+    // tests below still use it to derive plausibility, and Deserialize backs the #364 corpus structs.
+    use better_fleet::capture::is_plausible_sot_port;
+    use serde::Deserialize;
 
     // Real capture from issue #364 (in game): two flows, BOTH in the SoT port
     // range, but the real server carries 1247 packets vs 8. Volume must decide.
